Improve Tables
Added inline divs around tables on the clutter types page which allows the tables to sit next to each other. Removed value and mass from those tables to better fit the tables and clean the view overall. Used pymdownx.extra for markdown="1" to enabled markdown tables inside of the inlineDivs.